requirements.txt
# 核心框架
langgraph>=0.0.20
langchain>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 前端
streamlit>=1.28.0
plotly>=5.17.0

# 数据库
sqlalchemy>=2.0.0

# 配置
python-dotenv>=1.0.0
pyyaml>=6.0

# 工具
pydantic>=2.0.0